Remove the location step and just defer to the last location step
